Move DI container and Logging to before Avalonia
This is to enable headless operation.
